% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{as_igraph}
\alias{as_igraph}
\title{Given a mitre network it returns the same as igraph}
\usage{
as_igraph(mitrenet = getLatestDataSet()[["mitrenet"]], verbose = FALSE)
}
\arguments{
\item{mitrenet}{MITRE network built with this package}

\item{verbose}{default is FALSE}
}
\value{
igraph
}
\description{
Given a mitre network it returns the same as igraph
}
