% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attck.groups}
\alias{attck.groups}
\title{ATT&CK Groups Objects.}
\format{
A data frame with 11 variables.
}
\usage{
attck.groups
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
