% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attck.mitigations}
\alias{attck.mitigations}
\title{ATT&CK Mitigation Objects.}
\format{
A data frame with 12 variables.
}
\usage{
attck.mitigations
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
