% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attck}
\alias{attck}
\title{MITRE ATT&CK framework data set}
\format{
A list of 5 data frames and one visNetwork object:
\describe{
  \item{tactics}{a data frame with 28 rows and 13 columns.}
  \item{techniques}{a data frame with 611 rows and 16 columns.}
  \item{groups}{a data frame with 111 rows and 13 columns.}
  \item{software}{a data frame with 513 rows and 14 columns.}
  \item{mitigation}{a data frame with 55 rows and 14 columns.}
  \item{attcknet}{a visNetwork with 1314 nodes and 10058 edges.}
}
}
\source{
\url{https://github.com/mitre/cti}
}
\usage{
attck
}
\description{
A data set containing ATT&CK object models as data frames. It also contains
a graph network with the relations between objects.
}
\keyword{datasets}
