% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attck.relations}
\alias{attck.relations}
\title{ATT&CK relations Objects.}
\format{
A data frame with 11 variables.
}
\usage{
attck.relations
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
