% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attck.software}
\alias{attck.software}
\title{ATT&CK software Objects.}
\format{
A data frame with 12 variables.
}
\usage{
attck.software
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
