% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attck.techniques}
\alias{attck.techniques}
\title{ATT&CK techniques Objects.}
\format{
A data frame with 15 variables.
}
\usage{
attck.techniques
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
