% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.attck.R
\name{attck_latest_data}
\alias{attck_latest_data}
\title{Load ATT&CK data frame from local file or download latest}
\usage{
attck_latest_data(local_path = NA, remote = F, verbose = F)
}
\arguments{
\item{local_path}{path to RDS file. NA value implies remote TRUE}

\item{remote}{logical}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Load ATT&CK data frame from local file or download latest
}
