% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.R
\name{build_edges}
\alias{build_edges}
\title{Extract relationships between standards as edges in a data frame.}
\usage{
build_edges()
}
\value{
data.frame
}
\description{
\code{from} : node id of edge start
\code{to} : node id of edge end
\code{from_std} : standard id of edge start
\code{to_std} : standard id of edge end
\code{value} : When a value is set, the nodes will be scaled using the options in the scaling object defined above.
\code{title} : The title is shown in a pop-up when the mouse moves over the edge.
\code{arrows} : To draw an arrow with default settings a string can be supplied. For example: 'to, from,middle' or 'to;from', any combination with any separating symbol is fine. If you want to control the size of the arrowheads, you can supply an object.
\code{dashes} : When true, the edge will be drawn as a dashed line.
\code{color} : Color for the node.
\code{hidden} : When true, the node will not be shown. It will still be part of the physics simulation though!
}
