% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.R
\name{build_network}
\alias{build_network}
\title{Create a list of nodes and edges related to all standards in data folder.}
\usage{
build_network()
}
\value{
list, containing nodes and edges as data frames
}
\description{
Create a list of nodes and edges related to all standards in data folder.
}
