% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{buildAttckTactics}
\alias{buildAttckTactics}
\title{Parse tactics}
\usage{
buildAttckTactics(verbose = FALSE, deprecated = FALSE)
}
\arguments{
\item{verbose}{Default set as FALSE}

\item{deprecated}{Default set as FALSE, change it if you want to include deprecated objects}
}
\value{
data frame
}
\description{
Parse tactics
}
