% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capec.categories}
\alias{capec.categories}
\title{CAPEC categories Objects.}
\format{
A data frame with 11 variables.
}
\usage{
capec.categories
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
