% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capec.patterns}
\alias{capec.patterns}
\title{CAPEC patterns Objects.}
\format{
A data frame with 16 variables.
}
\usage{
capec.patterns
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
