% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capec.relations}
\alias{capec.relations}
\title{CAPEC relations Objects.}
\format{
A data frame with 4 variables.
}
\usage{
capec.relations
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
