% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{capec.views}
\alias{capec.views}
\title{CAPEC views Objects.}
\format{
A data frame with 11 variables.
}
\usage{
capec.views
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
