% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car.analytics}
\alias{car.analytics}
\title{CAR analytics Objects.}
\format{
A data frame with 11 variables.
}
\usage{
car.analytics
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
