% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car.coverage}
\alias{car.coverage}
\title{CAR coverage Objects.}
\format{
A data frame with 4 variables.
}
\usage{
car.coverage
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
