% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car.implementations}
\alias{car.implementations}
\title{CAR implementations Objects.}
\format{
A data frame with 7 variables.
}
\usage{
car.implementations
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
