% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{car.relations}
\alias{car.relations}
\title{CAR relations Objects.}
\format{
A data frame with 11 variables.
}
\usage{
car.relations
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
