% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.car.R
\name{car_latest_data}
\alias{car_latest_data}
\title{Load CAR data frame from local file or download latest}
\usage{
car_latest_data(local_path = NA, remote = F, verbose = F)
}
\arguments{
\item{local_path}{path to RDS file. NA value implies remote TRUE}

\item{remote}{logical}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Load CAR data frame from local file or download latest
}
