% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpe.nist}
\alias{cpe.nist}
\title{Common Platform Enumeration.}
\format{
A data frame with 16 variables: \code{title}, \code{cpe.22},
  \code{cpe.23}, and all separated values.
}
\usage{
cpe.nist
}
\description{
Full data set provided by NIST.
}
\keyword{datasets}
