% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_add_features}
\alias{cpe_add_features}
\title{Title}
\usage{
cpe_add_features(df = cpe.nist, scale_log = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{scale_log}{logical}
}
\value{
data.frame
}
\description{
Title
}
