% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_add_notation}
\alias{cpe_add_notation}
\title{Used by generic annotate cpes function}
\usage{
cpe_add_notation(
  df = cpe_latest_data(),
  type = c("vpv", "vp", "pv", "vv", "vend", "prod", "vers")[1]
)
}
\arguments{
\item{df}{data.frame}

\item{type}{character}
}
\value{
data.frame
}
\description{
Used by generic annotate cpes function
}
