% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cpe_find_vulnerabilities}
\alias{cpe_find_vulnerabilities}
\title{Title}
\usage{
cpe_find_vulnerabilities(
  df_inventory = cpe_make_title(),
  cves = cve_latest_data(),
  verbose = FALSE
)
}
\arguments{
\item{df_inventory}{data.frame}

\item{cves}{data.frame}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Title
}
