% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cpe.R
\name{cpe_generate}
\alias{cpe_generate}
\title{Title}
\usage{
cpe_generate(df = getInventory(), verbose = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Title
}
