% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_local_inventory}
\alias{cpe_local_inventory}
\title{Title}
\usage{
cpe_local_inventory(df = getInventory(), verbose = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Title
}
