% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_lstm_dataset}
\alias{cpe_lstm_dataset}
\title{Title}
\usage{
cpe_lstm_dataset(
  df = cpe_latest_data(),
  keep_deprecated = FALSE,
  only_wfn = TRUE
)
}
\arguments{
\item{df}{data.frame}

\item{keep_deprecated}{logical}

\item{only_wfn}{logical}
}
\value{
data.frame
}
\description{
Title
}
