% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{cpe_make_title}
\alias{cpe_make_title}
\title{Title}
\usage{
cpe_make_title(df_inventory = df, verbose = verbose)
}
\arguments{
\item{df_inventory}{data.frame}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Title
}
