% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_pca_stats}
\alias{cpe_pca_stats}
\title{Title}
\usage{
cpe_pca_stats(df = cpe_add_features(), verbose = FALSE, acum_prop = 0.75)
}
\arguments{
\item{df}{data.frame}

\item{verbose}{logical}

\item{acum_prop}{numeric}
}
\value{
data.frame
}
\description{
Title
}
