% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_sccm_inventory}
\alias{cpe_sccm_inventory}
\title{Title}
\usage{
cpe_sccm_inventory(
  path_sccm = "inst/extdata/sccm_component_definitions.csv",
  df_sccm = data.frame(),
  verbose = FALSE,
  headr = TRUE
)
}
\arguments{
\item{path_sccm}{path to sccm component definitions csv file}

\item{df_sccm}{data.frame}

\item{verbose}{logical}

\item{headr}{logical}
}
\value{
data.frame
}
\description{
Title
}
