% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_stats}
\alias{cpe_stats}
\title{Returns latest Official CPE as data frame grouped by vendor and product.}
\usage{
cpe_stats(df = cpe_latest_data(), only_vendor = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{only_vendor}{logical}
}
\value{
data.frame
}
\description{
Returns latest Official CPE as data frame grouped by vendor and product.
}
