% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cpe.R
\name{cpe_str2wfn}
\alias{cpe_str2wfn}
\title{Transform common string to WFN escaped name}
\usage{
cpe_str2wfn(name = c("Notepad++ v8.50", "CTS Projects & Software ClassAd 3.0"))
}
\arguments{
\item{name}{character vector}
}
\value{
character
}
\description{
Transform common string to WFN escaped name
}
