% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{cpe_train_set}
\alias{cpe_train_set}
\title{Title}
\usage{
cpe_train_set(cpes = mitre::cpe.nist, num_samples = 1000)
}
\arguments{
\item{cpes}{data.frame}

\item{num_samples}{integer}
}
\value{
data.frame
}
\description{
Title
}
