% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_valid_chars}
\alias{cpe_valid_chars}
\title{Returns a custom list of valid chars defined in notebook Explore_CPE_charset.
The parameters are used to customize the output.}
\usage{
cpe_valid_chars(
  taste = c("char", "dec", "hex")[1],
  add_tab = FALSE,
  add_enter = FALSE,
  add_underline = FALSE,
  type = c("input", "output")[1]
)
}
\arguments{
\item{taste}{character, type of output "char", "dec", or "hex"}

\item{add_tab}{logical, include tab char}

\item{add_enter}{logical, include enter char}

\item{add_underline}{logical, include "_"}

\item{type}{character, returns valid chars for "input" or "output"}
}
\value{
character
}
\description{
Returns a custom list of valid chars defined in notebook Explore_CPE_charset.
The parameters are used to customize the output.
}
