% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cpe.R
\name{cpe_wfn2str}
\alias{cpe_wfn2str}
\title{Transform WFN escaped name to common string
Ref: NIST IR 7695 (Aug. 2011) section 6.1.3.2, function decode}
\usage{
cpe_wfn2str(name = character())
}
\arguments{
\item{name}{character vector}
}
\value{
character
}
\description{
This function scans the string s and returns a copy
with all percent-encoded characters decoded. This
function is the inverse of pct_encode(s) defined in
Section 6.1.2.3. Only legal percent-encoded forms are
decoded. Others raise an error.
Decode a blank to logical ANY, and hyphen to logical NA.
}
