% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cpe.R
\name{cpe_wfn_add_quoting}
\alias{cpe_wfn_add_quoting}
\title{Inspect each character in string s. Copy quoted characters,
with their escaping, into the result. Look for unquoted non
alphanumerics and if not "*" or "?", add escaping.}
\usage{
cpe_wfn_add_quoting(s = character())
}
\arguments{
\item{s}{character}
}
\value{
character
}
\description{
REF: NIST IR 7695 (Aug. 2011) section 6.2.3.2
}
