% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_wfn_dataset}
\alias{cpe_wfn_dataset}
\title{Encode CPE columns ready for LSTM input.
Title is encoded to 73 valid characters.
Vendor, product and version are encoded to 49 valid chars.
If any character encoding fails, the CPE is filtered.}
\usage{
cpe_wfn_dataset(df = cpe_latest_data())
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
Encode CPE columns ready for LSTM input.
Title is encoded to 73 valid characters.
Vendor, product and version are encoded to 49 valid chars.
If any character encoding fails, the CPE is filtered.
}
