% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_wfn_process_quoted_chars}
\alias{cpe_wfn_process_quoted_chars}
\title{Inspect each character in string s. Certain nonalpha
characters pass thru without escaping into the result,
but most retain escaping}
\usage{
cpe_wfn_process_quoted_chars(s = character())
}
\arguments{
\item{s}{character}
}
\value{
character
}
\description{
REF: NIST IR 7695 (Aug. 2011) section 6.2.2.2
}
