% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cpe.R
\name{cpelite_vulnerable_configs}
\alias{cpelite_vulnerable_configs}
\title{Find vulnerabilities related to a CPE defined as cpelite (vendor:product) and version}
\usage{
cpelite_vulnerable_configs(
  x,
  x_vers,
  cves = cve_latest_data(),
  verbose = FALSE
)
}
\arguments{
\item{x}{character cpelite as vendor:product}

\item{x_vers}{character cpe version}

\item{cves}{data.frame}

\item{verbose}{logical}
}
\value{
character
}
\description{
Find vulnerabilities related to a CPE defined as cpelite (vendor:product) and version
}
