% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attck.R
\name{createATTCKedges}
\alias{createATTCKedges}
\title{Create edges from ATTCK data frames}
\usage{
createATTCKedges(
  tactics,
  techniques,
  mitigations,
  groups,
  software,
  relations,
  verbose
)
}
\arguments{
\item{tactics}{data.frame}

\item{techniques}{data.frame}

\item{mitigations}{data.frame}

\item{groups}{data.frame}

\item{software}{data.frame}

\item{relations}{data.frame}

\item{verbose}{Default set as FALSE}
}
\value{
data.frame
}
\description{
Create edges from ATTCK data frames
}
