% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{createTrainSample}
\alias{createTrainSample}
\title{Title}
\usage{
createTrainSample(
  cpes = data.frame(),
  num_samples = 5000,
  samples_mix = c(min = 0.6, med = 0.25, max = 0.15)
)
}
\arguments{
\item{cpes}{data.frame}

\item{num_samples}{integer}

\item{samples_mix}{numeric}
}
\value{
data.frame
}
\description{
Title
}
