% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cve.nist}
\alias{cve.nist}
\title{Common Vulnerability Enumeration.}
\format{
A data frame with 34 variables: \code{cve.id}, \code{problem.type} which is related to CWE,
  \code{description}, \code{vulnerable.configuration} which is related to CPE,
  \code{references}, \code{cvss3}, \code{cvss2} and all separated values.
}
\usage{
cve.nist
}
\description{
Full data set provided by NIST.
}
\keyword{datasets}
