% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cve.R
\name{cve_latest_data}
\alias{cve_latest_data}
\title{Load CVE data frame from local file or download latest}
\usage{
cve_latest_data(local_path = NA, remote = F, allcols = F, verbose = F)
}
\arguments{
\item{local_path}{path to RDS file. NA value implies remote TRUE}

\item{remote}{logical}

\item{allcols}{logical}

\item{verbose}{logical}

\item{keep_deprecated}{logical}
}
\value{
data.frame
}
\description{
Load CVE data frame from local file or download latest
}
