% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cwe.categories}
\alias{cwe.categories}
\title{CWE categories Objects.}
\format{
A data frame with 7 variables.
}
\usage{
cwe.categories
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
