% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cwe.weaknesses}
\alias{cwe.weaknesses}
\title{CWE Weaknesses Objects.}
\format{
A data frame with 24 variables.
}
\usage{
cwe.weaknesses
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
