% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cwe.R
\name{cwe_latest_data}
\alias{cwe_latest_data}
\title{Load CWE data frame from local file or download latest}
\usage{
cwe_latest_data(local_path = NA, remote = F, verbose = F)
}
\arguments{
\item{local_path}{path to RDS file. NA value implies remote TRUE}

\item{remote}{logical}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Load CWE data frame from local file or download latest
}
