% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.R
\name{downloadRawData}
\alias{downloadRawData}
\title{Download from official sources raw files saving them in [working_directory]/data-raw/}
\usage{
downloadRawData(verbose = FALSE)
}
\arguments{
\item{verbose}{default is FALSE}
}
\description{
Download from official sources raw files saving them in [working_directory]/data-raw/
}
\examples{
\donttest{
mitre::downloadRawData(T)
}
}
