% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.engage.R
\name{engage_latest_data}
\alias{engage_latest_data}
\title{Load ENGAGE data frame from local file or download latest}
\usage{
engage_latest_data(local_path = NA, remote = F, verbose = F)
}
\arguments{
\item{local_path}{path to RDS file. NA value implies remote TRUE}

\item{remote}{logical}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Load ENGAGE data frame from local file or download latest
}
