% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{etShieldTechniquesDetail}
\alias{etShieldTechniquesDetail}
\title{MITRE Shield Techniques detailed data frame}
\usage{
etShieldTechniquesDetail()
}
\value{
data.frame
}
\description{
MITRE Shield Techniques detailed data frame
}
