% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attck.R
\name{getAttckData}
\alias{getAttckData}
\title{ETL process that download current attck definitions and return a list of
data frames for each object. The list also contains a visNetwork object with
ATT&CK objects as nodes and all relations as edges.}
\usage{
getAttckData(verbose = FALSE)
}
\arguments{
\item{verbose}{Default set as FALSE}
}
\value{
list of data frames
}
\description{
ETL process that download current attck definitions and return a list of
data frames for each object. The list also contains a visNetwork object with
ATT&CK objects as nodes and all relations as edges.
}
\examples{
\dontrun{
attck <- getAttckData()
attck_tactics <- attck[["tactics"]]
attck_techniques <- attck[["techniques"]]
attck_groups <- attck[["groups"]]
attck_software <- attck[["software"]]
attck_mitigation <- attck[["mitigation"]]
}
}
