% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capec.R
\name{getCAPECData}
\alias{getCAPECData}
<<<<<<< HEAD
\title{Title}
=======
\title{ETL process that download current CAPEC definitions and return a list with a
data frame for CAPEC objects. The list also contains a visNetwork object with
CAPEC objects as nodes and all relations as edges.}
>>>>>>> devel
\usage{
getCAPECData(verbose = FALSE)
}
\arguments{
<<<<<<< HEAD
\item{verbose}{}
}
\value{

}
\description{
Title
=======
\item{verbose}{Default set as FALSE}
}
\value{
list of data frames
}
\description{
ETL process that download current CAPEC definitions and return a list with a
data frame for CAPEC objects. The list also contains a visNetwork object with
CAPEC objects as nodes and all relations as edges.
>>>>>>> devel
}
