% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpe.R
\name{getCPEData}
\alias{getCPEData}
\title{ETL process that download current CPE definitions and return a list with a
data frame for CPE objects. The list also contains a list with
CPE objects as nodes and all relations as edges.}
\usage{
getCPEData(
  verbose = FALSE,
  savepath = "data-raw/official-cpe-dictionary_v2.3.xml"
)
}
\arguments{
\item{verbose}{Default set as FALSE}

\item{savepath}{file path to store CPE dictionary}
}
\value{
list of data frames
}
\description{
ETL process that download current CPE definitions and return a list with a
data frame for CPE objects. The list also contains a list with
CPE objects as nodes and all relations as edges.
}
