% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{getCPEstats}
\alias{getCPEstats}
\title{Returns data frame with grouped count by vendor and product.}
\usage{
getCPEstats(only_vendor = TRUE, as_WFN = TRUE)
}
\arguments{
\item{only_vendor}{logical}

\item{as_WFN}{logical}
}
\value{
data.frame
}
\description{
Returns data frame with grouped count by vendor and product.
}
