% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attck.R
\name{getCurrentCTIdata}
\alias{getCurrentCTIdata}
\title{Retrieve current CTI definitions from github, and return data sets as list of
data frames: tactics, techniques, ...}
\usage{
getCurrentCTIdata()
}
\value{
ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .
  The data frame columns are: id, name, description, long_description.
}
\description{
Retrieve current CTI definitions from github, and return data sets as list of
data frames: tactics, techniques, ...
}
\examples{
\dontrun{
attck <- getCurrentCTIdata()
}
}
