% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{getGitHubCTIfiles}
\alias{getGitHubCTIfiles}
\title{Function to provide source JSON files by domain and STIX Object type}
\usage{
getGitHubCTIfiles(
  domain = sample(c("pre-attack", "enterprise-attack", "mobile-attack", "ics-attack"),
    1),
  object = sample(c("attack-pattern", "intrusion-set", "malware", "tool",
    "course-of-action", "x-mitre-tactic", "x-mitre-matrix"), 1)
)
}
\arguments{
\item{domain}{default set as random between "pre-attack", "enterprise-attack", "mobile-attack". Must be the same name of MITRE CTI Repository folders at github.com}

\item{object}{default set as random between "attack-pattern", "intrusion-set", "malware", "tool", "course-of-action", "x-mitre-tactic", "x-mitre-matrix". Must be the same name of MITRE CTI Repository folders at github.com}
}
\value{
data.frame with filename and raw url as columns
}
\description{
Function to provide source JSON files by domain and STIX Object type
}
\examples{
\dontrun{
pre.attck.pattern <- getGitHubCTIfiles(domain = "pre-attack", object = "attack-pattern")
mob.malware <- getGitHubCTIfiles(domain = "mob-attack", object = "malware")
}
}
