% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getInventary}
\alias{getInventary}
\title{Return data.frame with installed software name, version and vendor.}
\usage{
getInventary()
}
\value{
data.frame
}
\description{
Return data.frame with installed software name, version and vendor.
}
\examples{
inventory <- getInventary()
}
