% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getInventory}
\alias{getInventory}
\title{Return data.frame with installed software name, version and vendor.
Set parameter cpes as TRUE to predict CPEs using ML.}
\usage{
getInventory(verbose = FALSE, predict_cpes = FALSE)
}
\arguments{
\item{verbose}{logical}

\item{predict_cpes}{logical}
}
\value{
data.frame
}
\description{
Return data.frame with installed software name, version and vendor.
Set parameter cpes as TRUE to predict CPEs using ML.
}
\examples{
inventory <- getInventory()
}
