% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.R
\name{getMitreNetwork}
\alias{getMitreNetwork}
\title{Nodes and Edges ready for digraphs. Include shield and att&ck objects.}
\usage{
getMitreNetwork(verbose = FALSE)
}
\arguments{
\item{verbose}{default is FALSE}
}
\value{
list of two data frames: nodes and edges
}
\description{
Nodes and Edges ready for digraphs. Include shield and att&ck objects.
}
\examples{
\donttest{
mitrenet <- mitre::getMitreNetwork()
}
}
