% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{getNodeNeighbors}
\alias{getNodeNeighbors}
\title{Returns a list of nodes and edges (neighbors) based on input node.}
\usage{
getNodeNeighbors(
  node = "T1104",
  direction = "both",
  mitrenet = getLatestDataSet()[["mitrenet"]],
  verbose = FALSE
)
}
\arguments{
\item{node}{MITRE Standard Id}

\item{direction}{value should be: "from", "to" or "both"}

\item{mitrenet}{MITRE network built with this package}

\item{verbose}{default is FALSE}
}
\value{

}
\description{
Returns a list of nodes and edges (neighbors) based on input node.
}
