% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldData}
\alias{getShieldData}
\title{ETL process that download current shield definitions and return a list of
data frames for each object. The list also contains a visNetwork object with
SHIELD objects as nodes and all relations as edges.}
\usage{
getShieldData(verbose = FALSE)
}
\arguments{
\item{verbose}{Default set as FALSE}
}
\value{
list of data frames
}
\description{
ETL process that download current shield definitions and return a list of
data frames for each object. The list also contains a visNetwork object with
SHIELD objects as nodes and all relations as edges.
}
