% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldNodes}
\alias{getShieldNodes}
\title{MITRE Shield objects as nodes in a data frame}
\usage{
getShieldNodes()
}
\value{
data.frame
}
\description{
MITRE Shield objects as nodes in a data frame
}
\keyword{internal}
