% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldOpportunities}
\alias{getShieldOpportunities}
\title{MITRE Shield Opportunities data frame}
\usage{
getShieldOpportunities()
}
\value{
ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .
  The data frame columns are: id, description.
}
\description{
MITRE Shield Opportunities data frame
}
\examples{
\donttest{
opportunities <- getShieldOpportunities()
}
}
