% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldProcedures}
\alias{getShieldProcedures}
\title{ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .}
\usage{
getShieldProcedures()
}
\value{
MITRE Shield Procedures data frame
  The data frame columns are: id, description.
}
\description{
ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .
}
