% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldRelations}
\alias{getShieldRelations}
\title{MITRE Shield objects relations data frame}
\usage{
getShieldRelations()
}
\value{
data.frame
}
\description{
MITRE Shield objects relations data frame
}
\keyword{internal}
