% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldTactics}
\alias{getShieldTactics}
\title{ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .}
\usage{
getShieldTactics()
}
\value{
MITRE Shield Tactics data frame
  The data frame columns are: id, name, description, long_description.
}
\description{
ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .
}
