% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldTactictDetail}
\alias{getShieldTactictDetail}
\title{MITRE Shield Tactics detailed data frame}
\usage{
getShieldTactictDetail()
}
\value{
data.frame
}
\description{
MITRE Shield Tactics detailed data frame
}
