% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldTechniquesDetail}
\alias{getShieldTechniquesDetail}
\title{MITRE Shield Techniques detailed data frame}
\usage{
getShieldTechniquesDetail()
}
\value{
data.frame
}
\description{
MITRE Shield Techniques detailed data frame
}
