% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldUseCases}
\alias{getShieldUseCases}
\title{ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .}
\usage{
getShieldUseCases()
}
\value{
MITRE Shield Use Cases data frame
  The data frame columns are: id, description.
}
\description{
ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .
}
\keyword{internal}
