% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{MapCommonproperties}
\alias{MapCommonproperties}
\title{Extract common properties from attack pattern object (parsed with RJSONIO::fromJSON)}
\usage{
MapCommonproperties(attack.obj = NA, domain = NA)
}
\arguments{
\item{attack.obj}{list}

\item{domain}{must be "pre-attack", "enterprise-attack" or "mobile-attack"}
}
\value{
data.frame compliant with CTI USAGE document
}
\description{
Extract common properties from attack pattern object (parsed with RJSONIO::fromJSON)
}
\examples{
\dontrun{
sf <- "https://github.com/mitre/cti/raw/master/<domain>/<object>/<file>.json"
attack.pattern <- RJSONIO::fromJSON(sf)
df.common <- MapCommonproperties(attack.pattern)
}
}
