% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{MapGroups}
\alias{MapGroups}
\title{Extract Group properties from intrusion set object (parsed with RJSONIO::fromJSON)}
\usage{
MapGroups(intrusion.set = NA, domain = NA)
}
\arguments{
\item{intrusion.set}{list based on STIX}

\item{domain}{must be "pre-attack", "enterprise-attack" or "mobile-attack"}
}
\value{
data.frame
}
\description{
Extract Group properties from intrusion set object (parsed with RJSONIO::fromJSON)
}
