% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{MapMitigation}
\alias{MapMitigation}
\title{Extract Mitigation properties from course.action object (parsed with RJSONIO::fromJSON)}
\usage{
MapMitigation(course.action = NA, domain = domain)
}
\arguments{
\item{course.action}{list based on STIX}

\item{domain}{must be "pre-attack", "enterprise-attack" or "mobile-attack"}
}
\value{
data.frame
}
\description{
Extract Mitigation properties from course.action object (parsed with RJSONIO::fromJSON)
}
