% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{MapRelations}
\alias{MapRelations}
\title{Extract object relationships from relationship object (parsed with RJSONIO::fromJSON)}
\usage{
MapRelations(relationship = NA, domain = NA)
}
\arguments{
\item{relationship}{list based on STIX}

\item{domain}{must be "pre-attack", "enterprise-attack" or "mobile-attack"}
}
\value{
data.frame
}
\description{
Extract object relationships from relationship object (parsed with RJSONIO::fromJSON)
}
