% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{MapSoftware}
\alias{MapSoftware}
\title{Extract Software properties from malware and tool object (parsed with RJSONIO::fromJSON)}
\usage{
MapSoftware(software.obj = NA, domain = domain)
}
\arguments{
\item{software.obj}{list based on STIX}

\item{domain}{must be "pre-attack", "enterprise-attack" or "mobile-attack"}
}
\value{
data.frame
}
\description{
Extract Software properties from malware and tool object (parsed with RJSONIO::fromJSON)
}
