% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml_cpe_valid_chars}
\alias{ml_cpe_valid_chars}
\title{Title}
\usage{
ml_cpe_valid_chars(
  taste = c("char", "dec", "hex")[1],
  add_tab = FALSE,
  add_enter = FALSE
)
}
\arguments{
\item{add_enter}{}
}
\description{
Title
}
