% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ner_offsets_cpe}
\alias{ner_offsets_cpe}
\title{Title}
\usage{
ner_offsets_cpe(cpes = mitre::cpe.nist, overlap = TRUE, columns = "all")
}
\arguments{
\item{columns}{}
}
\description{
Title
}
\examples{
\dontrun{
 df <- ner_offsets_cpe()
 write.csv(df[, c("title", "annotations")], "inst/extdata/cpes_notoverlap.csv",
           row.names = FALSE, fileEncoding = "UTF-8")
}
}
