% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ner_tags_cpe}
\alias{ner_tags_cpe}
\title{Title}
\usage{
ner_tags_cpe(cpes = mitre::cpe.nist, overlap = TRUE, columns = "all")
}
\arguments{
\item{cpes}{data.frame}

\item{overlap}{logical, default TRUE}

\item{columns}{character}
}
\value{
data.frame
}
\description{
Title
}
\examples{
\dontrun{
 df <- ner_tags_cpe()
}
}
