% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{nlp.ner_cpe_trainset}
\alias{nlp.ner_cpe_trainset}
\title{Title}
\usage{
nlp.ner_cpe_trainset(
  num_samples = 5000,
  type = c("vpv", "vp", "pv", "vv", "vend", "prod", "vers")[1],
  kind = c("RASA", "entities")[1],
  pydict = TRUE,
  rdataset = "data-raw/NLP/cpes_vpv_rasa.rds",
  seed = 42
)
}
\arguments{
\item{num_samples}{integer}

\item{type}{character, default vpv}

\item{kind}{character, default RASA}

\item{pydict}{logical, python offsets begins with 0, otherwise 1}

\item{rdataset}{character, path to RDS. nlp_cpe_annotate default if not exists}

\item{seed}{integer, used for reproducible research}
}
\value{
data.frame
}
\description{
Title
}
