% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{nlp_cpe_annotate}
\alias{nlp_cpe_annotate}
\title{Create annotations for NER into new column `annotation`. By default, it uses
RASA notation and detect vendor, product and version.
Rows with overlapped entities are skipped.
The output can be configured with parameters:
 - `type` to select which entities annotate
 - `kind` for notation: RASA, BILUO, entities}
\usage{
nlp_cpe_annotate(
  df = nlp_cpe_dataset(),
  type = c("vpv", "vp", "pv", "vv", "vend", "prod", "version")[1],
  kind = c("RASA", "entities", "BILUO")[1],
  pydict = TRUE
)
}
\arguments{
\item{df}{data.frame nlp_cpe_dataset}

\item{type}{character, default vpv}

\item{kind}{character, default RASA}

\item{pydict}{logical, python offsets begins with 0, otherwise 1}
}
\value{
data.frame
}
\description{
Create annotations for NER into new column `annotation`. By default, it uses
RASA notation and detect vendor, product and version.
Rows with overlapped entities are skipped.
The output can be configured with parameters:
 - `type` to select which entities annotate
 - `kind` for notation: RASA, BILUO, entities
}
