% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{nlp_cpe_build_ner_trainset}
\alias{nlp_cpe_build_ner_trainset}
\title{Title}
\usage{
nlp_cpe_build_ner_trainset(
  seed = 42,
  num_samples = 10000,
  mix_config = list(keep_deprecated = FALSE, only_wfn = FALSE, scale_log = FALSE,
    sample_weight = c("pca", "vendor", "none")[1], pca_features = c("len_vendor",
    "abc_vendor", "num_version", "len_version", "abc_product"), vendor_qntl = c(0, 0.8,
    0.9, 0.99, 1)),
  ner_config = list(notation = c("rasa", "offset")[1], vendor = TRUE, product = TRUE,
    version = TRUE),
  save_path = file.path("C:", "DEVEL", "code", "data", "ner_trainsets"),
  verbose = FALSE
)
}
\arguments{
\item{seed}{integer}

\item{num_samples}{integer}

\item{mix_config}{list}

\item{ner_config}{list}

\item{save_path}{character}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
Title
}
