% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{nlp_cpe_dataset}
\alias{nlp_cpe_dataset}
\title{Remove rows containing non valid characters for CPE LSTM tokenization}
\usage{
nlp_cpe_dataset(df = mitre::cpe.nist)
}
\arguments{
\item{df}{data.frame default mitre cpes}
}
\value{
data.frame
}
\description{
Remove rows containing non valid characters for CPE LSTM tokenization
}
