% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{nlp_cpe_feateng}
\alias{nlp_cpe_feateng}
\title{Title}
\usage{
nlp_cpe_feateng(df = nlp_cpe_dataset(), scale_log = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{scale_log}{logical}
}
\value{
data.frame
}
\description{
Title
}
