% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{nlp_cpe_rasa_notation}
\alias{nlp_cpe_rasa_notation}
\title{Used by generic annotate cpes function}
\usage{
nlp_cpe_rasa_notation(
  df_ner = data.frame(),
  type = c("vpv", "vp", "pv", "vv", "vend", "prod", "version")[1]
)
}
\arguments{
\item{df_ner}{data.frame}

\item{type}{character}
}
\value{
data.frame
}
\description{
Used by generic annotate cpes function
}
