% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{nlp_cpe_sample_dataset}
\alias{nlp_cpe_sample_dataset}
\title{Title}
\usage{
nlp_cpe_sample_dataset(
  df = nlp_cpe_dataset(),
  num_samples = 1000,
  seed = 42,
  quantiles = c(0, 0.8, 0.9, 0.99, 1)
)
}
\arguments{
\item{df}{data.frame}

\item{num_samples}{integer}

\item{seed}{integer}

\item{quantiles}{numeric}
}
\value{
data.frame
}
\description{
Title
}
