% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{omitDeprecated}
\alias{omitDeprecated}
\title{Given a mitre network it returns the same without deprecated nodes}
\usage{
omitDeprecated(mitrenet = getLatestDataSet()[["mitrenet"]], verbose = FALSE)
}
\arguments{
\item{mitrenet}{MITRE network built with this package}

\item{verbose}{default is FALSE}
}
\value{
list of nodes and edges
}
\description{
Given a mitre network it returns the same without deprecated nodes
}
