% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{parseAttck.Groups}
\alias{parseAttck.Groups}
\title{Read MITRE CTI Repository browsing domain directories to extract data from intrusion-set files,
map variables from STIX to ATT&CK model and return tidy data.frame with Group variables.}
\usage{
parseAttck.Groups(verbose = TRUE)
}
\arguments{
\item{verbose}{default is FALSE}
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository browsing domain directories to extract data from intrusion-set files,
map variables from STIX to ATT&CK model and return tidy data.frame with Group variables.
}
\examples{
\dontrun{
df.groups <- parseAttck.Groups()
}
}
