% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{parseAttck.Relationships}
\alias{parseAttck.Relationships}
\title{Read MITRE CTI Repository browsing domain directories to extract data from relationship files,
build model and return tidy data.frame with relationship variables.}
\usage{
parseAttck.Relationships(verbose = TRUE)
}
\arguments{
\item{verbose}{default is FALSE}
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository browsing domain directories to extract data from relationship files,
build model and return tidy data.frame with relationship variables.
}
\examples{
\dontrun{
df.relationships <- parseAttck.Relationships()
}
}
