% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{parseAttck.Software}
\alias{parseAttck.Software}
\title{Read MITRE CTI Repository browsing domain directories to extract data from malware and tool files,
build model and return tidy data.frame with Software variables.}
\usage{
parseAttck.Software(verbose = TRUE)
}
\arguments{
\item{verbose}{default is FALSE}
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository browsing domain directories to extract data from malware and tool files,
build model and return tidy data.frame with Software variables.
}
\examples{
\dontrun{
df.software <- parseAttck.Software()
}
}
