% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{parseAttck.Tactics}
\alias{parseAttck.Tactics}
\title{Read MITRE CTI Repository browsing domain directories to extract data from x-mitre-tactic files,
map variables from STIX to ATT&CK model and return tidy data.frame with Tactic variables.}
\usage{
parseAttck.Tactics(verbose = TRUE)
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository browsing domain directories to extract data from x-mitre-tactic files,
map variables from STIX to ATT&CK model and return tidy data.frame with Tactic variables.
}
\examples{
\dontrun{
df.tactics <- parseAttck.Tactics()
}
}
