% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attck.R
\name{parseAttckData}
\alias{parseAttckData}
\title{ETL process that download current attck definitions and return a list of
data frames for each object. The list also contains a visNetwork object with
ATT&CK objects as nodes and all relations as edges.}
\usage{
parseAttckData(verbose = FALSE)
}
\arguments{
\item{verbose}{Default set as FALSE}
}
\value{
list of data frames
}
\description{
ETL process that download current attck definitions and return a list of
data frames for each object. The list also contains a visNetwork object with
ATT&CK objects as nodes and all relations as edges.
}
\examples{
\donttest{
attck <- mitre::parseAttckData()
}
}
