% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{parseAttckmodel.soft}
\alias{parseAttckmodel.soft}
\title{Read MITRE CTI Repository files related to malware and tool, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.}
\usage{
parseAttckmodel.soft(
  domain = sample(c("pre-attack", "ics-attack", "enterprise-attack", "mobile-attack"),
    1),
  verbose = TRUE
)
}
\arguments{
\item{domain}{must be "pre-attack", "enterprise-attack" or "mobile-attack"}

\item{verbose}{default is FALSE}
}
\value{
data.frame
}
\description{
Read MITRE CTI Repository files related to malware and tool, extract data,
map variables from STIX to ATT&CK model and return tidy data.frame.
}
