% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cti.R
\name{parseCTIdata}
\alias{parseCTIdata}
\title{Parse ATT&CK source files from MITRE CTI Repository and build raw data sets}
\usage{
parseCTIdata(verbose = TRUE)
}
\value{
list of data.frame objects (tactics, techniques, groups, software and relationships)
}
\description{
Parse ATT&CK source files from MITRE CTI Repository and build raw data sets
}
