% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.R
\name{parseRawData}
\alias{parseRawData}
\title{ETL process for all standards, it also create a list of nodes and edges
representing the relationships between standard objects. It needs raw files
pre-downloaded from official MITRE repositories stored in a folder named "data-raw".}
\usage{
parseRawData(verbose = FALSE)
}
\arguments{
\item{verbose}{default is TRUE}
}
\value{
list of two data frames: nodes and edges
}
\description{
ETL process for all standards, it also create a list of nodes and edges
representing the relationships between standard objects. It needs raw files
pre-downloaded from official MITRE repositories stored in a folder named "data-raw".
}
\examples{
\dontrun{
mitredata <- mitre::parseRawData(TRUE)
}
}
