% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cpe.R
\name{predict_cpe}
\alias{predict_cpe}
\title{Create a CPE for each inventory item.}
\usage{
predict_cpe(
  df_inventory = mitre::getInventory(),
  model_name = "Neurona/cpener-test"
)
}
\arguments{
\item{df_inventory}{data.frame with columns: vendor, name and version; title is optional}

\item{model_name}{huggingface reference, by default: Neurona/cpener-test}
}
\value{
data.frame
}
\description{
Create a CPE for each inventory item.
}
