% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shield.procedures}
\alias{shield.procedures}
\title{SHIELD procedures Objects.}
\format{
A data frame with 11 variables.
}
\usage{
shield.procedures
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
