% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shield}
\alias{shield}
\title{MITRE SHIELD framework dataset}
\format{
A list of 5 data frames and one visNetwork object:
\describe{
  \item{tactics}{a data frame with 8 rows and 4 columns.}
  \item{techniques}{a data frame with 33 rows and 4 columns.}
  \item{opportunities}{a data frame with 79 rows and 2 columns.}
  \item{procedures}{a data frame with 66 rows and 2 columns.}
  \item{usecases}{a data frame with 189 rows and 2 columns.}
  \item{shieldnet}{a visNetwork with 375 nodes and 835 edges.}
}
}
\source{
\url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data}
}
\usage{
shield
}
\description{
A dataset containing SHIELD object models as data frames. It also contains
a graph network with the relations between objects.
}
\keyword{datasets}
