% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shield.relations}
\alias{shield.relations}
\title{SHIELD relations Objects.}
\format{
A data frame with 3 variables.
}
\usage{
shield.relations
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
