% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shield.tactics}
\alias{shield.tactics}
\title{SHIELD tactics Objects.}
\format{
A data frame with 4 variables.
}
\usage{
shield.tactics
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
