% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shield.techniques}
\alias{shield.techniques}
\title{SHIELD techniques Objects.}
\format{
A data frame with 4 variables.
}
\usage{
shield.techniques
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
