% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shield.use_cases}
\alias{shield.use_cases}
\title{SHIELD use cases Objects.}
\format{
A data frame with 2 variables.
}
\usage{
shield.use_cases
}
\description{
Full data set provided by MITRE
}
\keyword{datasets}
