% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cpe.R
\name{str49enc}
\alias{str49enc}
\title{Encode strings to only 49 accepted characters for custom WFN.}
\usage{
str49enc(name = character(), na_replace = "*")
}
\arguments{
\item{name}{character vector with CPE names}

\item{na_replace}{character, no valid chars will be replaced with "*" by default}
}
\value{
character
}
\description{
Encode strings to only 49 accepted characters for custom WFN.
}
