% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cpe.R
\name{str73enc}
\alias{str73enc}
\title{Encode strings to only 73 accepted characters for custom WFN.
  - Replace accents, dieresis, etc to simple ASCII chars
  - Replace tabs with spaces
  - Deal with valid escaped symbols
  - Replace not valid characters with "*"}
\usage{
str73enc(name = character(), na_replace = "*")
}
\arguments{
\item{name}{character vector with CPE names}

\item{na_replace}{character, no valid chars will be replaced with "*" by default}
}
\value{
character
}
\description{
Encode strings to only 73 accepted characters for custom WFN.
  - Replace accents, dieresis, etc to simple ASCII chars
  - Replace tabs with spaces
  - Deal with valid escaped symbols
  - Replace not valid characters with "*"
}
