% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.R
\name{updateRawData}
\alias{updateRawData}
\title{Download from official sources raw files saving them in [package_path]/data-raw/}
\usage{
updateRawData(verbose = FALSE)
}
\arguments{
\item{verbose}{default is FALSE}
}
\description{
Download from official sources raw files saving them in [package_path]/data-raw/
}
