% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{wfn2str}
\alias{wfn2str}
\title{Transform WFN escaped name to common string}
\usage{
wfn2str(name = character())
}
\arguments{
\item{name}{character vector}
}
\value{
character
}
\description{
Transform WFN escaped name to common string
}
