% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.R
\name{build_network}
\alias{build_network}
\title{Create a list of nodes and edges related to all standards in data folder.}
\usage{
build_network(verbose = FALSE, as_igraph = TRUE)
}
\arguments{
\item{verbose}{logical, FALSE by default. Change it to see the process messages.}

\item{as_igraph}{logical, TRUE by default. Change it to get list of nodes and edges.}
}
\value{
list, containing nodes and edges as data frames
}
\description{
Create a list of nodes and edges related to all standards in data folder.
}
\examples{
\dontrun{
mitrenet <- mitre::build_network(as_igraph = FALSE)
}
}
