% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cpe.R
\name{cpe_generate}
\alias{cpe_generate}
\title{This function complements the inventory data.frame generating CPEs using
heuristics and NER transformer. It adds columns with predicted CPE, the
scoring and possible CPE candidates.}
\usage{
cpe_generate(df_inventory = getInventory(), verbose = FALSE)
}
\arguments{
\item{df_inventory}{data.frame}

\item{verbose}{logical}
}
\value{
data.frame
}
\description{
This function complements the inventory data.frame generating CPEs using
heuristics and NER transformer. It adds columns with predicted CPE, the
scoring and possible CPE candidates.
}
