% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_latest_data}
\alias{cpe_latest_data}
\title{Load CPE data frame from local file or download latest}
\usage{
cpe_latest_data(local_path = NA, remote = F)
}
\arguments{
\item{local_path}{path to RDS file. NA value implies remote TRUE}

\item{remote}{logical}
}
\value{
data.frame
}
\description{
Load CPE data frame from local file or download latest
}
\examples{
cpes <- mitre::cpe_latest_date(local_path = "inst/extdata/cpe.nist.rds")
}
