% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_lstm_dataset}
\alias{cpe_lstm_dataset}
\title{Prepare CPE data frame valid for LSTM training.}
\usage{
cpe_lstm_dataset(df = cpe_latest_data(), verbose = F)
}
\arguments{
\item{df}{data.frame}

\item{verbose}{}
}
\value{
data.frame
}
\description{
Prepare CPE data frame valid for LSTM training.
}
