% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp_cpe.R
\name{cpe_stats}
\alias{cpe_stats}
\title{Returns data frame with grouped count by vendor and product.}
\usage{
cpe_stats(df = data.frame(), only_vendor = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{only_vendor}{logical}
}
\value{
data.frame
}
\description{
Returns data frame with grouped count by vendor and product.
}
