% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpe.R
\name{getCPEData}
\alias{getCPEData}
\title{ETL process that download current CPE definitions and return a list with a
data frame for CPE objects. The list also contains a visNetwork object with
CPE objects as nodes and all relations as edges.}
\usage{
getCPEData(verbose = FALSE)
}
\arguments{
\item{verbose}{Default set as FALSE}
}
\value{
list of data frames
}
\description{
ETL process that download current CPE definitions and return a list with a
data frame for CPE objects. The list also contains a visNetwork object with
CPE objects as nodes and all relations as edges.
}
