% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.cpe.R
\name{getCPEstats}
\alias{getCPEstats}
\title{Returns data frame with grouped count by vendor and product.}
\usage{
getCPEstats(df = cpe_latest_data(), only_vendor = TRUE, scale_log = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{only_vendor}{logical}

\item{scale_log}{logical}
}
\value{
data.frame
}
\description{
Returns data frame with grouped count by vendor and product.
}
