% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwe.R
\name{getCWEData}
\alias{getCWEData}
\title{ETL process that download current CVE definitions and return a list with a
data frame for CVE objects. The list also contains a visNetwork object with
CVE objects as nodes and all relations as edges.}
\usage{
getCWEData(cwes.file = "data-raw/cwec_v4.3.xml", verbose = FALSE)
}
\arguments{
\item{cwes.file}{raw xml file}

\item{verbose}{Default set as FALSE}
}
\value{
data frame
}
\description{
ETL process that download current CVE definitions and return a list with a
data frame for CVE objects. The list also contains a visNetwork object with
CVE objects as nodes and all relations as edges.
}
\examples{
\donttest{
cwes <- mitre::getCWEData()
}
}
