% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getInventory}
\alias{getInventory}
\title{Return data.frame with installed software name, version and vendor.}
\usage{
getInventory()
}
\value{
data.frame
}
\description{
Return data.frame with installed software name, version and vendor.
}
\examples{
inventory <- getInventory()
}
