% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getInventory}
\alias{getInventory}
\title{Return data.frame with installed software name, version and vendor.
Set predict_cpes as TRUE to predict CPEs using ML.
Set predict_cves as TRUE to predict CPEs and its vulnerabilities as CVEs}
\usage{
getInventory(
  include_libs = FALSE,
  verbose = FALSE,
  predict_cpes = FALSE,
  predict_cves = FALSE
)
}
\arguments{
\item{include_libs}{logical}

\item{verbose}{logical}

\item{predict_cpes}{logical}

\item{predict_cves}{logical}
}
\value{
data.frame
}
\description{
Return data.frame with installed software name, version and vendor.
Set predict_cpes as TRUE to predict CPEs using ML.
Set predict_cves as TRUE to predict CPEs and its vulnerabilities as CVEs
}
\examples{
inventory <- getInventory()
}
