% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldNetwork}
\alias{getShieldNetwork}
\title{MITRE  ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .
It returns a visNetwork object ready for analyze and plot.}
\usage{
getShieldNetwork()
}
\value{
list of nodes and edges
}
\description{
MITRE  ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .
It returns a visNetwork object ready for analyze and plot.
}
\examples{
\donttest{
shieldnet <- getShieldNetwork()
}
}
\keyword{internal}
