% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldOpportunities}
\alias{getShieldOpportunities}
\title{MITRE Shield Opportunities data frame}
\usage{
getShieldOpportunities()
}
\value{
data.frame
}
\description{
MITRE Shield Opportunities data frame
}
