% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldTactics}
\alias{getShieldTactics}
\title{MITRE Shield Tactics data frame}
\usage{
getShieldTactics()
}
\value{
ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .
  The data frame columns are: id, name, description, long_description.
}
\description{
MITRE Shield Tactics data frame
}
\examples{
\donttest{
shield.tactics <- getShieldTactics()
}
}
