% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldTechniques}
\alias{getShieldTechniques}
\title{MITRE Shield Techniques data frame}
\usage{
getShieldTechniques()
}
\value{
ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data} .
  The data frame columns are: id, name, description, long_description.
}
\description{
MITRE Shield Techniques data frame
}
\examples{
\donttest{
shield.techniques <- getShieldTechniques()
}
}
