% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shield.R
\name{getShieldUseCases}
\alias{getShieldUseCases}
\title{MITRE Shield Use Cases data frame}
\usage{
getShieldUseCases()
}
\value{
ETL process that read source data from \url{https://github.com/MITRECND/mitrecnd.github.io/tree/master/_data}
  The data frame columns are: id, description.
}
\description{
MITRE Shield Use Cases data frame
}
\examples{
\dontrun{
usecases <- getShieldUseCases()
}
}
