% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlp.R
\name{nlp_cpe_dataset}
\alias{nlp_cpe_dataset}
\title{Transform the CPEs data frame removing unnecessary columns. Change default parameters
to remove deprecated or those that doesn't follow the WFN definition.}
\usage{
nlp_cpe_dataset(df = mitre::cpe.nist, keep_deprecated = FALSE, only_wfn = TRUE)
}
\arguments{
\item{df}{data.frame default mitre cpes}

\item{keep_deprecated}{logical, condition to keep or remove deprecated CPEs}

\item{only_wfn}{logical, condition to keep or remove CPEs following WFN definition}
}
\value{
data.frame
}
\description{
Transform the CPEs data frame removing unnecessary columns. Change default parameters
to remove deprecated or those that doesn't follow the WFN definition.
}
