% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitre.R
\name{parseRawData}
\alias{parseRawData}
\title{ETL process for all standards, it also create a list of nodes and edges
representing the relationships between standard objects. It needs raw files
downloaded from official MITRE repositories stored in a folder named "data-raw";
set downloadLatest parameter to TRUE and the function will create it for you.}
\usage{
parseRawData(verbose = FALSE, downloadLatest = TRUE)
}
\arguments{
\item{verbose}{default is FALSE}

\item{downloadLatest}{default as FALSE, set to TRUE to download latest raw source files from official MITRE repositories}
}
\value{
list of two data frames: nodes and edges
}
\description{
ETL process for all standards, it also create a list of nodes and edges
representing the relationships between standard objects. It needs raw files
downloaded from official MITRE repositories stored in a folder named "data-raw";
set downloadLatest parameter to TRUE and the function will create it for you.
}
\examples{
\donttest{
mitredata <- mitre::parseRawData(T)
}
}
